Finished coding the backend can make further changes later

+++ urbandictionerybackend/src/controller/DictionaryUserController.ts
@@ -6,10 +6,10 @@ import { Route } from '../decorator/Route'
 import { validate, ValidationError, ValidatorOptions } from 'class-validator'
 import { UrbanTermDefinition } from '../entity/UrbanTermDefinition'
 
-@Controller('/Users')
+@Controller('/user')
 export default class DictionaryUserController {
-  private readonly userRepo = AppDataSource.getRepository(DictionaryUser) // Student Repository
-  private readonly defRepo = AppDataSource.getRepository(UrbanTermDefinition) // Student Repository
+  private readonly userRepo = AppDataSource.getRepository(DictionaryUser) // User Repository
+  private readonly defRepo = AppDataSource.getRepository(UrbanTermDefinition) // Definition Repository
 
   // https://github.com/typestack/class-validator#passing-options
   private readonly validOptions: ValidatorOptions = {
@@ -23,28 +23,39 @@ export default class DictionaryUserController {
 
   @Route('get', '/:id')
   async read (req: Request, res: Response, next: NextFunction): Promise<DictionaryUser> {
-    if (req.body) {
-      const userToCheck = await this.userRepo.findOneBy({ id: req.params.id })
-      if (userToCheck.Password.localeCompare(req.body.Password) === 0) {
-        res.status = 201
-        return userToCheck
+    if (req.body && req.body.id === req.params.id) {
+      const userToCheck = await this.userRepo.findOne({
+        relations: { definitions: true },
+        where: { id: req.params.id }
+      })
+      if (userToCheck) {
+        const password = await this.userRepo.findOne({ select: { Password: true }, where: { id: req.params.id } })
+        if (password.Password === req.body.Password) {
+          res.status = 201
+          return userToCheck
+        } else {
+          res.status = 401
+          return req.body
+        }
       } else {
         res.status = 401
         return req.body
       }
+    } else {
+      res.status = 401
+      return req.body
     }
-
-    return await this.userRepo.findOneBy({ id: req.params.id })
   }
 
   @Route('delete', '/:id')
   async delete (req: Request, res: Response, next: NextFunction): Promise<DictionaryUser> {
-    const userToRemove = await this.userRepo.findOneBy({ id: req.params.id })
+    const userToRemove = await this.userRepo.findOne({ relations: { definitions: true }, where: { id: req.params.id } })
     res.status = 204
     if (userToRemove) {
-      const definitionsToRemove = await this.defRepo.findBy({ UserID: userToRemove.id })
-      for (const def of definitionsToRemove) {
-        await this.defRepo.remove(def)
+      if (userToRemove.definitions) {
+        for (const def of userToRemove.definitions) {
+          await this.defRepo.remove(def)
+        }
       }
       return await this.userRepo.remove(userToRemove)
     } else next()
@@ -59,7 +70,7 @@ export default class DictionaryUserController {
     else {
       const violations = await validate(userToUpdate, this.validOptions)
       if (violations.length) {
-        res.statusCode = 422 // Unprocessable Entity
+        res.status = 422 // Unprocessable Entity
         return violations
       } else {
         return await this.userRepo.save(userToUpdate)
@@ -73,10 +84,10 @@ export default class DictionaryUserController {
     const newUser = Object.assign(new DictionaryUser(), req.body)
     const violations = await validate(newUser, this.validOptions)
     if (violations.length) {
-      res.statusCode = 422 // Unprocessable Entity
+      res.status = 422 // Unprocessable Entity
       return violations
     } else {
-      res.statusCode = 201
+      res.status = 201
       return await this.userRepo.save(newUser)
     }
   }
